op-revm,op-reth: report zero DA footprint on PostExec receipts
Jovian repurposes the receipt's `blobGasUsed` field to carry the transaction's
DA footprint. op-reth's RPC receipt builder recomputed that value from the tx
bytes for every transaction once Jovian was active, with no type gate, so a
PostExec (`0x7D`) tx reported a non-zero footprint — the Fjord minimum-size
floor guarantees at least `100 * daFootprintGasScalar`, never zero.

The block-level accumulator in alloy-op-evm already excludes the `0x7D`, and
`blobGasUsed` is not a consensus receipt field, so the executor's zero never
reaches the RPC layer. Gate the recomputation on the tx type instead, restoring
`sum(receipt.blobGasUsed) == header.blobGasUsed`.

Extract `da_footprint_gas(encoded, scalar)` in op-revm as
`encoded_tx_da_footprint` (plus a `tx_da_footprint` convenience wrapper that
encodes first), next to `estimate_tx_compressed_size`: the arithmetic was
duplicated in the block executor, the RPC receipt builder, and two tests. The
`0x7D` check lives in the helper and keys off the first byte of the EIP-2718
encoding, so it needs no trait bound beyond `Encodable2718` and no new method on
the transaction types. Free functions rather than methods — op-revm depends on
op-alloy-consensus (not the reverse), so a trait method could not reach the FLZ
estimator, and a `&self` method would force a re-encode where the block executor
reuses `tx_env.encoded_bytes()`.

Deposits are still excluded by their call sites, not by the helper: the block
total excludes deposits while their receipts report a non-zero footprint.
specs#926 does not cover that discrepancy and changing it is a user-visible RPC
change, so it is left alone.

The receipt test runs on a Lagoon-active schedule, since a `0x7D` at a
Jovian-only timestamp is not a reachable state — SDM rides Lagoon. No named
chain schedules Lagoon yet, hence the hand-built fork list, which appends a
single entry rather than restating the tail of the schedule: `OpChainHardforks`
resolves an activation positionally (`forks[fork.idx()]`), so a duplicate row
shifts every later fork by one. Two assertions pin that Karst's activation is
unchanged.

Also adds an end-to-end pin in alloy-op-evm that a `0x7D` leaves
`da_footprint_used` unchanged, and an SDM acceptance assertion that per-tx
receipt footprints sum to the header total on a real Lagoon block — the only
in-tree chain whose blocks contain a `0x7D`. The Jovian acceptance suites pin
Jovian at genesis, so Lagoon (and therefore SDM) can never be active there.

The `Encodable2718` bound means op-revm picks up a new `alloy-eips` dependency,
which accounts for the otherwise-unrelated build churn: the `serde` and `std`
features have to propagate to it (zepter), and the three satellite lockfiles
that path-depend on op-revm — `kona/sp1/programs`, `op-rbuilder`,
`rollup-boost` — each record its dependency list, so all three needed
refreshing. They were regenerated without `--locked`, so only the moved entry
changed and no unrelated dep was bumped.

Specification: https://github.com/ethereum-optimism/specs/pull/926
