+++ rust/op-revm/src/lib.rs
@@ -29,4 +29,7 @@ pub use evm::OpEvm;
 pub use l1block::L1BlockInfo;
 pub use result::OpHaltReason;
 pub use spec::*;
-pub use transaction::{OpTransaction, error::OpTransactionError, estimate_tx_compressed_size};
+pub use transaction::{
+    OpTransaction, encoded_tx_da_footprint, error::OpTransactionError, estimate_tx_compressed_size,
+    tx_da_footprint,
+};
